updated for version 7.0031

+++ src/gui_kde_wid.cc
@@ -1540,7 +1540,7 @@ SBPool::create(GuiScrollbar * sb, int orient)//{{{
     }
 
     connect(sb->w, SIGNAL(valueChanged(int)), &mapper, SLOT(map()));
-    mapper.setMapping(sb->w, (int)sb);
+    mapper.setMapping(sb->w, (long)sb);
 }//}}}
